requirements.txt
pyyaml
numpy
pandas
blinker
networkx
PyGithub